plans added in page/search.js and lib/search.js
