+++ petup/lib/search.js
@@ -3,7 +3,8 @@
 
 //we can eitherfetch exernal API database or query database directly. 
 
-import { user } from '../db/sampledata'
+import { user } from '../db/sampledata' // this is old dummy data. will refactor to show newer data
+
 
 //the function in this folder will query to get data from database
 //for now, the data {user} is stored locally and being imported in this file
@@ -15,6 +16,63 @@ export function getAllUsers() {
     // return res.json();
     return res
 }
+
+function getSitterByCity(){
+    //get service by city. do direct sql query to database
+    const response = [];//SQLquery here
+    return response;
+    //should it be return response.json()?
+
+    //for loop function to find matching key for and returns an array of user object that matches key city
+  function citySearch(response, input) {
+    let result = [];
+    for (let i = 0; i < response.length; i++) {
+      if (response[i].address.city == input) {
+        result.push(response[i]);
+      }
+    }
+    console.log(`this is results array: ${result}`);
+
+//16/1 - will update the function above
+    //!!!!!!12/1: props has changed. displayresults card needs to be updated with new prop data
+
+    //each user object is like this: 
+    // {id: 1,
+    // name: "Leanne Graham",
+    // nickname: "Leanne",
+    // imageURL : "https://img.freepik.com/free-photo/pleasant-looking-serious-man-stands-profile-has-confident-expression-wears-casual-white-t-shirt_273609-16959.jpg?w=2000",
+    // tagline: "I am a pet carer",
+    // address: {
+    //   street: "Kulas Light",
+    //   city: "London",
+    //   postcode: "WN5 9FT",
+    // },
+    // phone: "1-770-736-8031 x56442",
+    // rate : "35"}
+
+    return result;
+  }
+
+
+}
+
+function getSitterByCityServicePets(){
+    //get service by city and type of pets (1 filter)
+    // 3. get service by city, type of service, type of pet 
+}
+
+
+export {getSitterByCity,
+    getSitterByCityServicePets}
+
+//todo 16.1 : 
+//set up a function that gives data based on the filter below:
+//thefunctions that we need (for filter): 
+// 1. get all service by city(for utmost basic search, no filter)
+// 2. get service by city and type of pets (1 filter)
+// 3. get service by city, type of service, type of pet  (2 and 3 can be combined)
+// 4. get service by city, type of service, type of pet, price range (min and max)
+// 5. get service by city, type of service, type of pet, proce range, date of availability
 
 //sample codes to get data from database below:
